feat(zudo-doc): ship gen-z-index + tags-audit as package bins (#2334)
S9b — ship the z-index codegen and tags-audit tooling as bins of
@takazudo/zudo-doc, keeping all source/DATA project-side.

- gen-z-index bin: pure-Node ESM, reads src/config/z-index-tokens.ts and
  rewrites the GENERATED:Z_INDEX block in src/styles/global.css from the
  project root (process.cwd()); --check exits 1 on drift.
- tags-audit bin: spawns the project's tsx with a runner that dynamically
  imports the project's TypeScript config (settings.ts, tag-vocabulary.ts)
  and runs the shared core from @takazudo/zudo-doc/tags-audit.
- New @takazudo/zudo-doc/tags-audit export: core audit library compiled from
  src/tags-audit.ts; pluralize/string-similarity decoupled via NearDupHelpers.
- package.json: APPEND bin + ./tags-audit export per S2 convention; add bin to
  files array.
- Re-point project + generated scripts: gen:z-index, check:z-index, tags:audit
  now invoke the package bins.
- scripts/tags-audit.ts becomes a thin re-export shim (backward-compat for
  scripts/__tests__/tags-audit.test.ts); template kept byte-identical.
- Docs: src/CLAUDE.md (z-index tooling) + src/config/CLAUDE.md (tags-audit
  ownership) describe the package-side tooling / project-side data split.
