+++ package.json
@@ -45,8 +45,8 @@
     "setup:zudo-doc-wisdom": "bash scripts/setup-zudo-doc-wisdom.sh",
     "// ── Linting ─────────────────────────────────────": "",
     "lint:tokens": "design-token-lint",
-    "gen:z-index": "node scripts/gen-z-index.mjs",
-    "check:z-index": "node scripts/gen-z-index.mjs --check",
+    "gen:z-index": "gen-z-index",
+    "check:z-index": "gen-z-index --check",
     "check:template-drift": "bash scripts/check-template-drift.sh",
     "check:fixture-settings-drift": "node scripts/check-fixture-settings-drift.mjs",
     "check:pin-parity": "node scripts/check-pin-parity.mjs",
@@ -55,7 +55,7 @@
     "check:flaky-tracking-issue": "node scripts/check-flaky-tracking-issue.mjs",
     "check:package-safelist": "node scripts/check-package-safelist.mjs",
     "// ── Tags ────────────────────────────────────────": "",
-    "tags:audit": "tsx scripts/tags-audit.ts",
+    "tags:audit": "tags-audit",
     "tags:suggest": "tsx scripts/tags-suggest.ts",
     "// ── Lifecycle (auto-run by pnpm install) ─────────": "",
     "prepare": "lefthook install --reset-hooks-path && bash scripts/install-git-hooks.sh",
